add bloodType enum

+++ src/main/java/com/keyrus/kit/models/Dna.java
@@ -2,19 +2,19 @@ package com.keyrus.kit.models;
 
 public class Dna {
 
-    private Integer id;
+    private Long id;
     private String dna;
 
-    public Dna(Integer id, String dna) {
+    public Dna(Long id, String dna) {
         this.id = id;
         this.dna = dna;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
